requirements.txt
opencv-python>=4.8.0
ultralytics>=8.0.0  # YOLOv8
numpy>=1.24.0
PyYAML>=6.0.1
click>=8.1.0  # for CLI and progress bars
python-dotenv>=1.0.0  # for environment variables
python-dateutil>=2.8.2  # for robust date parsing and comparison 